put fork setup on tracing executor
