+++ evm/src/trace/mod.rs
@@ -1,7 +1,7 @@
 use crate::{
     abi::CHEATCODE_ADDRESS,
     debug::Instruction,
-    executor::{fork::CreateFork, Backend, Executor, ExecutorBuilder},
+    executor::{fork::CreateFork, opts::EvmOpts, Backend, Executor, ExecutorBuilder},
     trace::identifier::LocalTraceIdentifier,
     utils::evm_spec,
     CallKind,
@@ -14,9 +14,13 @@ use ethers::{
     types::{Bytes, DefaultFrame, GethDebugTracingOptions, StructLog, H256, U256},
 };
 use foundry_common::contracts::{ContractsByAddress, ContractsByArtifact};
+use foundry_config::{find_project_root_path, Config};
 use hashbrown::HashMap;
 use node::CallTraceNode;
-use revm::interpreter::{opcode, CallContext, InstructionResult, Memory, Stack};
+use revm::{
+    interpreter::{opcode, CallContext, InstructionResult, Memory, Stack},
+    primitives::Env,
+};
 use serde::{Deserialize, Serialize};
 use std::{
     collections::{BTreeMap, HashSet},
@@ -59,6 +63,20 @@ impl TracingExecutor {
         executor.set_tracing(true).set_debugger(debug);
 
         Self { executor }
+    }
+
+    pub async fn get_fork_material(
+        config: &Config,
+        mut evm_opts: EvmOpts,
+    ) -> eyre::Result<(Env, Option<CreateFork>, Option<ethers::types::Chain>)> {
+        evm_opts.fork_url = Some(config.get_rpc_url_or_localhost_http()?.into_owned());
+        evm_opts.fork_block_number = config.fork_block_number;
+
+        let env = evm_opts.evm_env().await;
+
+        let fork = evm_opts.get_fork(config, env.clone());
+
+        Ok((env, fork, evm_opts.get_remote_chain_id()))
     }
 }
